StageFright: fix parsing SPS frame.
Buffer size has been changed to off_t to catch and handle range errors.
Parsing has been fixed according the ITU-T H.264 (06/2011).

+++ include/media/stagefright/OMXCodec.h
@@ -71,7 +71,7 @@ class CBitBuffer {
     // Pointer to buffer
     const uint8_t *m_pBuffer;
     // Buffer size
-    size_t m_nBufferSize;
+    off_t m_nBufferSize;
     // Number of read bits
     uint8_t m_nBitsRead;
